refactor json conversion; simplify ParseMap via b64dec_bigint util
