+++ jwt-util/jwt_util/go/core/factory.go
@@ -50,41 +50,21 @@ func ParseJWK(jwk string) *rsa.PrivateKey {
 	return ParseMap(data)
 }
 
-func b64dec_bigint(data any) *big.Int {
-	data_str, ok := data.(string)
-	if !ok {
-		panic("input data is not a string")
-	}
-	buffer := b64dec(data_str)
-	output := new(big.Int)
-	output.SetBytes(buffer)
-	return output
-}
-
 func ParseMap(data StringMap) *rsa.PrivateKey {
-	//n := b64dec(data["n"].(string))
-	e := b64dec(data["e"].(string))
-	d := b64dec(data["d"].(string))
-	p := b64dec(data["p"].(string))
-	q := b64dec(data["q"].(string))
-	dp := b64dec(data["dp"].(string))
-	dq := b64dec(data["dq"].(string))
-	qi := b64dec(data["qi"].(string))
 	return &rsa.PrivateKey{
 		PublicKey: rsa.PublicKey{
-			//N: new(big.Int).SetBytes(n),
 			N: b64dec_bigint(data["n"]),
-			E: int(new(big.Int).SetBytes(e).Int64()),
+			E: int(b64dec_bigint(data["e"]).Int64()),
 		},
-		D: new(big.Int).SetBytes(d),
+		D: b64dec_bigint(data["d"]),
 		Primes: []*big.Int{
-			new(big.Int).SetBytes(p),
-			new(big.Int).SetBytes(q),
+			b64dec_bigint(data["p"]),
+			b64dec_bigint(data["q"]),
 		},
 		Precomputed: rsa.PrecomputedValues{
-			Dp:   new(big.Int).SetBytes(dp),
-			Dq:   new(big.Int).SetBytes(dq),
-			Qinv: new(big.Int).SetBytes(qi),
+			Dp:   b64dec_bigint(data["dp"]),
+			Dq:   b64dec_bigint(data["dq"]),
+			Qinv: b64dec_bigint(data["qi"]),
 		},
 	}
 }
